[server] Move snapshot service out of EE
