+++ components/server/ee/src/workspace/gitpod-server-impl.ts
@@ -74,7 +74,6 @@ import { StripeService } from "../user/stripe-service";
 import { GitHubAppSupport } from "../github/github-app-support";
 import { GitLabAppSupport } from "../gitlab/gitlab-app-support";
 import { Config } from "../../../src/config";
-import { SnapshotService, WaitForSnapshotOptions } from "./snapshot-service";
 import { ClientMetadata, traceClientMetadata } from "../../../src/websocket/websocket-connection-manager";
 import { BitbucketAppSupport } from "../bitbucket/bitbucket-app-support";
 import { URL } from "url";
@@ -106,8 +105,6 @@ export class GitpodServerEEImpl extends GitpodServerImpl {
     @inject(BitbucketAppSupport) protected readonly bitbucketAppSupport: BitbucketAppSupport;
 
     @inject(Config) protected readonly config: Config;
-
-    @inject(SnapshotService) protected readonly snapshotService: SnapshotService;
 
     @inject(UserService) protected readonly userService: UserService;
 
@@ -371,101 +368,6 @@ export class GitpodServerEEImpl extends GitpodServerImpl {
             workspace.shareable = level === "everyone";
             await db.store(workspace);
         });
-    }
-
-    async takeSnapshot(ctx: TraceContext, options: GitpodServer.TakeSnapshotOptions): Promise<string> {
-        traceAPIParams(ctx, { options });
-        const { workspaceId, dontWait } = options;
-        traceWI(ctx, { workspaceId });
-
-        const user = this.checkAndBlockUser("takeSnapshot");
-
-        const workspace = await this.guardSnaphotAccess(ctx, user.id, workspaceId);
-
-        const instance = await this.workspaceDb.trace(ctx).findRunningInstance(workspaceId);
-        if (!instance) {
-            throw new ResponseError(ErrorCodes.NOT_FOUND, `Workspace ${workspaceId} has no running instance`);
-        }
-        await this.guardAccess({ kind: "workspaceInstance", subject: instance, workspace }, "get");
-
-        const client = await this.workspaceManagerClientProvider.get(instance.region);
-        const request = new TakeSnapshotRequest();
-        request.setId(instance.id);
-        request.setReturnImmediately(true);
-
-        // this triggers the snapshots, but returns early! cmp. waitForSnapshot to wait for it's completion
-        const resp = await client.takeSnapshot(ctx, request);
-
-        const snapshot = await this.snapshotService.createSnapshot(options, resp.getUrl());
-
-        // to be backwards compatible during rollout, we require new clients to explicitly pass "dontWait: true"
-        const waitOpts = { workspaceOwner: workspace.ownerId, snapshot };
-        if (!dontWait) {
-            // this mimicks the old behavior: wait until the snapshot is through
-            await this.internalDoWaitForWorkspace(waitOpts);
-        } else {
-            // start driving the snapshot immediately
-            this.internalDoWaitForWorkspace(waitOpts).catch((err) =>
-                log.error({ userId: user.id, workspaceId: workspaceId }, "internalDoWaitForWorkspace", err),
-            );
-        }
-
-        return snapshot.id;
-    }
-
-    protected async guardSnaphotAccess(ctx: TraceContext, userId: string, workspaceId: string): Promise<Workspace> {
-        traceAPIParams(ctx, { userId, workspaceId });
-
-        const workspace = await this.workspaceDb.trace(ctx).findById(workspaceId);
-        if (!workspace || workspace.ownerId !== userId) {
-            throw new ResponseError(ErrorCodes.NOT_FOUND, `Workspace ${workspaceId} does not exist.`);
-        }
-        await this.guardAccess({ kind: "snapshot", subject: undefined, workspace }, "create");
-
-        return workspace;
-    }
-
-    /**
-     * @param snapshotId
-     * @throws ResponseError with either NOT_FOUND or SNAPSHOT_ERROR in case the snapshot is not done yet.
-     */
-    async waitForSnapshot(ctx: TraceContext, snapshotId: string): Promise<void> {
-        traceAPIParams(ctx, { snapshotId });
-
-        const user = this.checkAndBlockUser("waitForSnapshot");
-
-        const snapshot = await this.workspaceDb.trace(ctx).findSnapshotById(snapshotId);
-        if (!snapshot) {
-            throw new ResponseError(ErrorCodes.NOT_FOUND, `No snapshot with id '${snapshotId}' found.`);
-        }
-        const snapshotWorkspace = await this.guardSnaphotAccess(ctx, user.id, snapshot.originalWorkspaceId);
-        await this.internalDoWaitForWorkspace({ workspaceOwner: snapshotWorkspace.ownerId, snapshot });
-    }
-
-    protected async internalDoWaitForWorkspace(opts: WaitForSnapshotOptions) {
-        try {
-            await this.snapshotService.waitForSnapshot(opts);
-        } catch (err) {
-            // wrap in SNAPSHOT_ERROR to signal this call should not be retried.
-            throw new ResponseError(ErrorCodes.SNAPSHOT_ERROR, err.toString());
-        }
-    }
-
-    async getSnapshots(ctx: TraceContext, workspaceId: string): Promise<string[]> {
-        traceAPIParams(ctx, { workspaceId });
-        traceWI(ctx, { workspaceId });
-
-        const user = this.checkAndBlockUser("getSnapshots");
-
-        const workspace = await this.workspaceDb.trace(ctx).findById(workspaceId);
-        if (!workspace || workspace.ownerId !== user.id) {
-            throw new ResponseError(ErrorCodes.NOT_FOUND, `Workspace ${workspaceId} does not exist.`);
-        }
-
-        const snapshots = await this.workspaceDb.trace(ctx).findSnapshotsByWorkspaceId(workspaceId);
-        await Promise.all(snapshots.map((s) => this.guardAccess({ kind: "snapshot", subject: s, workspace }, "get")));
-
-        return snapshots.map((s) => s.id);
     }
 
     async adminGetUsers(ctx: TraceContext, req: AdminGetListRequest<User>): Promise<AdminGetListResult<User>> {
